refactor(api): split server.js into ES6 routers and lib modules
- server.js reduced from 2170 lines to 96 (entry point only)
- Fix stray require("net") to ES6 import

lib/ (shared utilities):
  - context.js: Winston logger, shared state (logs, localstackStatus), setIo, addLog
  - aws.js: execAsync, internalEndpoint, userEndpoint, awsRegion, awsEnv helper
  - upload.js: multer instance with 100MB limit
  - tcp.js: checkTcpPort utility
  - resources.js: createResources, createSingleResource, destroyResources, destroySingleResource, listResources, listAllBuckets
  - s3.js: listBucketContents
  - dynamodb.js: listDynamoDBTables, scanDynamoDBTable, queryDynamoDBTable, getDynamoDBTableSchema

routes/ (Express Router files):
  health.js, localstack.js, resources.js, config.js, profile.js,
  projects.js, savedConfigs.js, s3.js, dynamodb.js, cache.js,
  mailpit.js, secrets.js, postgres.js, keycloak.js
