+++ localcloud-api/lib/upload.js
@@ -1,0 +1,13 @@
+import multer from "multer";
+import fs from "fs";
+
+if (!fs.existsSync("/tmp/uploads")) {
+  fs.mkdirSync("/tmp/uploads", { recursive: true });
+}
+
+export const upload = multer({
+  dest: "/tmp/uploads/",
+  limits: {
+    fileSize: 100 * 1024 * 1024, // 100MB limit
+  },
+});
