+++ localcloud-api/server.js
@@ -1,18 +1,26 @@
 import express from "express";
 import cors from "cors";
-import { exec } from "child_process";
-import { promisify } from "util";
-import axios from "axios";
-import winston from "winston";
-import cron from "node-cron";
 import http from "http";
 import { Server as SocketIOServer } from "socket.io";
-import path from "path";
-import fs from "fs";
-import multer from "multer";
-import nodemailer from "nodemailer";
+import cron from "node-cron";
 
-const execAsync = promisify(exec);
+import { setIo, addLog } from "./lib/context.js";
+import { internalEndpoint } from "./lib/aws.js";
+
+import healthRouter from "./routes/health.js";
+import localstackRouter, { checkLocalStackStatus } from "./routes/localstack.js";
+import resourcesRouter from "./routes/resources.js";
+import configRouter from "./routes/config.js";
+import profileRouter from "./routes/profile.js";
+import projectsRouter from "./routes/projects.js";
+import savedConfigsRouter from "./routes/savedConfigs.js";
+import s3Router from "./routes/s3.js";
+import dynamodbRouter from "./routes/dynamodb.js";
+import cacheRouter from "./routes/cache.js";
+import mailpitRouter from "./routes/mailpit.js";
+import secretsRouter from "./routes/secrets.js";
+import postgresRouter from "./routes/postgres.js";
+import keycloakRouter from "./routes/keycloak.js";
 
 const app = express();
 const server = http.createServer(app);
@@ -25,54 +33,7 @@ const io = new SocketIOServer(server, {
   path: "/ws/socket.io",
 });
 
-// Configure logging
-const logger = winston.createLogger({
-  level: "info",
-  format: winston.format.combine(
-    winston.format.timestamp(),
-    winston.format.json()
-  ),
-  transports: [
-    new winston.transports.File({ filename: "logs/error.log", level: "error" }),
-    new winston.transports.File({ filename: "logs/combined.log" }),
-    new winston.transports.Console({
-      format: winston.format.combine(
-        winston.format.colorize(),
-        winston.format.simple()
-      ),
-    }),
-  ],
-});
-
-// In-memory storage for logs and status
-let logs = [];
-let localstackStatus = {
-  running: false,
-  endpoint: process.env.AWS_ENDPOINT_URL || "http://localhost:4566",
-  health: "unknown",
-  uptime: null,
-};
-
-// Internal endpoint for Docker networking (used by backend)
-const internalEndpoint = "http://localstack:4566";
-// User-friendly endpoint for GUI display
-const userEndpoint = "http://localhost:4566";
-const awsRegion = process.env.AWS_DEFAULT_REGION || "us-east-1";
-
-import db from "./db.js";
-
-// Configure multer for file uploads
-const upload = multer({
-  dest: "/tmp/uploads/",
-  limits: {
-    fileSize: 100 * 1024 * 1024, // 100MB limit
-  },
-});
-
-// Ensure upload directory exists
-if (!fs.existsSync("/tmp/uploads")) {
-  fs.mkdirSync("/tmp/uploads", { recursive: true });
-}
+setIo(io);
 
 // Middleware
 app.use(
@@ -81,1614 +42,28 @@ app.use(
     credentials: true,
   })
 );
-app.use(express.json({ limit: "50mb" })); // Increase limit for legacy JSON uploads
+app.use(express.json({ limit: "50mb" }));
 app.use(express.urlencoded({ extended: true, limit: "50mb" }));
 
-// Helper function to add log entry
-function addLog(level, message, source = "api") {
-  const logEntry = {
-    timestamp: new Date().toISOString(),
-    level,
-    message,
-    source,
-  };
-  logs.push(logEntry);
-
-  // Keep only last 1000 logs
-  if (logs.length > 1000) {
-    logs = logs.slice(-1000);
-  }
-
-  // Emit to connected clients
-  io.emit("log", logEntry);
-
-  logger.log(level, message, { source });
-}
-
-// LocalStack Management Functions
-async function checkLocalStackStatus() {
-  try {
-    addLog(
-      "info",
-      `Checking LocalStack health at ${internalEndpoint}`,
-      "localstack"
-    );
-    const response = await axios.get(`${internalEndpoint}/_localstack/health`, {
-      timeout: 5000,
-    });
-
-    if (response.status === 200) {
-      localstackStatus.running = true;
-      localstackStatus.health = "healthy";
-      addLog("info", "LocalStack is running and healthy", "localstack");
-    } else {
-      localstackStatus.running = false;
-      localstackStatus.health = "unhealthy";
-      addLog("warn", "LocalStack is running but unhealthy", "localstack");
-    }
-  } catch (error) {
-    localstackStatus.running = false;
-    localstackStatus.health = "unknown";
-    addLog(
-      "error",
-      `LocalStack health check failed: ${error.message}`,
-      "localstack"
-    );
-  }
-}
-
-// Resource Management Functions
-async function createResources(request) {
-  try {
-    const { projectName, resources, template } = request;
-
-    addLog("info", `Creating resources for ${projectName}`, "automation");
-
-    // Create resources individually for better control
-    const createdResources = [];
-    const errors = [];
-
-    // Create S3 bucket if requested
-    if (resources.s3) {
-      try {
-        const result = await createSingleResource(projectName, "s3");
-        createdResources.push(result);
-        addLog("success", `S3 bucket created: ${result.name}`, "automation");
-      } catch (error) {
-        errors.push(`S3: ${error.message}`);
-        addLog(
-          "error",
-          `Failed to create S3 bucket: ${error.message}`,
-          "automation"
-        );
-      }
-    }
-
-    // Create DynamoDB table if requested
-    if (resources.dynamodb) {
-      try {
-        const result = await createSingleResource(projectName, "dynamodb");
-        createdResources.push(result);
-        addLog(
-          "success",
-          `DynamoDB table created: ${result.name}`,
-          "automation"
-        );
-      } catch (error) {
-        errors.push(`DynamoDB: ${error.message}`);
-        addLog(
-          "error",
-          `Failed to create DynamoDB table: ${error.message}`,
-          "automation"
-        );
-      }
-    }
-
-    // Create Lambda function if requested
-    if (resources.lambda) {
-      try {
-        const result = await createSingleResource(projectName, "lambda");
-        createdResources.push(result);
-        addLog(
-          "success",
-          `Lambda function created: ${result.name}`,
-          "automation"
-        );
-      } catch (error) {
-        errors.push(`Lambda: ${error.message}`);
-        addLog(
-          "error",
-          `Failed to create Lambda function: ${error.message}`,
-          "automation"
-        );
-      }
-    }
-
-    // Create API Gateway if requested
-    if (resources.apigateway) {
-      try {
-        const result = await createSingleResource(projectName, "apigateway");
-        createdResources.push(result);
-        addLog("success", `API Gateway created: ${result.name}`, "automation");
-      } catch (error) {
-        errors.push(`API Gateway: ${error.message}`);
-        addLog(
-          "error",
-          `Failed to create API Gateway: ${error.message}`,
-          "automation"
-        );
-      }
-    }
-
-    // Create Secrets Manager secret if requested
-    if (resources.secretsmanager) {
-      try {
-        const result = await createSingleResource(
-          projectName,
-          "secretsmanager"
-        );
-        createdResources.push(result);
-        addLog(
-          "success",
-          `Secrets Manager secret created: ${result.name}`,
-          "automation"
-        );
-      } catch (error) {
-        errors.push(`Secrets Manager: ${error.message}`);
-        addLog(
-          "error",
-          `Failed to create Secrets Manager secret: ${error.message}`,
-          "automation"
-        );
-      }
-    }
-
-    if (errors.length > 0) {
-      addLog(
-        "warn",
-        `Some resources failed to create: ${errors.join(", ")}`,
-        "automation"
-      );
-      return {
-        success: createdResources.length > 0,
-        message: `Created ${
-          createdResources.length
-        } resources successfully. Errors: ${errors.join(", ")}`,
-        createdResources,
-        errors,
-      };
-    }
-
-    addLog(
-      "success",
-      `All ${createdResources.length} resources created successfully for ${projectName}`,
-      "automation"
-    );
-    return {
-      success: true,
-      message: `All ${createdResources.length} resources created successfully`,
-      createdResources,
-    };
-  } catch (error) {
-    addLog(
-      "error",
-      `Failed to create resources: ${error.message}`,
-      "automation"
-    );
-    return { success: false, error: error.message };
-  }
-}
-
-async function createSingleResource(projectName, resourceType, config = {}) {
-  try {
-    let command = `./create_single_resource.sh ${projectName} ${resourceType}`;
-
-    // For DynamoDB with configuration, pass the config as JSON
-    if (resourceType === "dynamodb" && config.dynamodbConfig) {
-      const configJson = JSON.stringify(config.dynamodbConfig);
-      command += ` --config '${configJson}'`;
-    }
-
-    // For S3 with configuration, pass the config as JSON
-    if (resourceType === "s3" && config.s3Config) {
-      const configJson = JSON.stringify(config.s3Config);
-      command += ` --config '${configJson}'`;
-    }
-
-    // For Secrets Manager with configuration, pass the config as JSON
-    if (resourceType === "secretsmanager" && config.secretsmanagerConfig) {
-      const configJson = JSON.stringify(config.secretsmanagerConfig);
-      command += ` --config '${configJson}'`;
-    }
-
-    // Add logging for debugging
-    console.log("[DEBUG] Running command:", command);
-
-    const { stdout, stderr } = await execAsync(command, {
-      cwd: "/app/scripts/shell",
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    // Log stdout and stderr
-    console.log("[DEBUG] Script stdout:", stdout);
-    console.log("[DEBUG] Script stderr:", stderr);
-
-    if (stderr) {
-      addLog(
-        "warn",
-        `Single resource creation warning: ${stderr}`,
-        "automation"
-      );
-    }
-
-    // Parse the output to get resource details
-    try {
-      const resourceInfo = JSON.parse(stdout);
-      return resourceInfo;
-    } catch (parseError) {
-      // If output is not JSON, create a basic resource info object
-      return {
-        id: `${resourceType}-${projectName}-${resourceType}`,
-        name: `${projectName}-${resourceType}`,
-        type: resourceType,
-        status: "active",
-        project: projectName,
-        createdAt: new Date().toISOString(),
-      };
-    }
-  } catch (error) {
-    throw new Error(`Failed to create ${resourceType}: ${error.message}`);
-  }
-}
-
-async function destroyResources(request) {
-  try {
-    const { projectName, resourceIds } = request;
-
-    addLog(
-      "info",
-      `Destroying specific resources for ${projectName}: ${
-        resourceIds ? resourceIds.join(", ") : "all resources"
-      }`,
-      "automation"
-    );
-
-    let command = `./destroy_resources.sh ${projectName} local`;
-
-    // Add specific resource IDs if provided
-    if (resourceIds && resourceIds.length > 0) {
-      command += ` ${resourceIds.join(" ")}`;
-    }
-
-    const { stdout, stderr } = await execAsync(command, {
-      cwd: "/app/scripts/shell",
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog("warn", `Resource destruction warning: ${stderr}`, "automation");
-    }
-
-    addLog(
-      "success",
-      `Resources destroyed successfully for ${projectName}`,
-      "automation"
-    );
-
-    return { success: true, message: "Resources destroyed successfully" };
-  } catch (error) {
-    addLog(
-      "error",
-      `Failed to destroy resources: ${error.message}`,
-      "automation"
-    );
-    return { success: false, error: error.message };
-  }
-}
-
-async function destroySingleResource(projectName, resourceType, resourceName) {
-  try {
-    let command = `./destroy_single_resource.sh ${projectName} ${resourceType} ${resourceName}`;
-
-    const { stdout, stderr } = await execAsync(command, {
-      cwd: "/app/scripts/shell",
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog("warn", `Resource destruction warning: ${stderr}`, "automation");
-    }
-
-    addLog(
-      "success",
-      `Resource destroyed successfully for ${projectName}`,
-      "automation"
-    );
-
-    return { success: true, message: "Resource destroyed successfully" };
-  } catch (error) {
-    addLog(
-      "error",
-      `Failed to destroy resource: ${error.message}`,
-      "automation"
-    );
-    return { success: false, error: error.message };
-  }
-}
-
-async function listResources(projectName) {
-  try {
-    let command = `./list_resources.sh ${projectName} local --all`;
-
-    const { stdout, stderr } = await execAsync(command, {
-      cwd: "/app/scripts/shell",
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog("warn", `Resource listing warning: ${stderr}`, "automation");
-    }
-
-    // Parse the output as JSON array
-    let resources = [];
-    try {
-      resources = JSON.parse(stdout);
-      if (!Array.isArray(resources)) {
-        addLog(
-          "error",
-          "Resource listing output is not a JSON array",
-          "automation"
-        );
-        resources = [];
-      }
-    } catch (err) {
-      addLog(
-        "error",
-        `Failed to parse resource listing JSON: ${err.message}`,
-        "automation"
-      );
-      resources = [];
-    }
-
-    return resources;
-  } catch (error) {
-    addLog("error", `Failed to list resources: ${error.message}`, "automation");
-    return [];
-  }
-}
-
-async function listAllBuckets(projectName) {
-  try {
-    let command = `./list_resources.sh ${projectName} local --all`;
-
-    const { stdout, stderr } = await execAsync(command, {
-      cwd: "/app/scripts/shell",
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog("warn", `Bucket listing warning: ${stderr}`, "automation");
-    }
-
-    // Parse the output as JSON array and filter for S3 buckets only
-    let allResources = [];
-    try {
-      allResources = JSON.parse(stdout);
-      if (!Array.isArray(allResources)) {
-        addLog(
-          "error",
-          "Resource listing output is not a JSON array",
-          "automation"
-        );
-        return [];
-      }
-    } catch (err) {
-      addLog(
-        "error",
-        `Failed to parse resource listing JSON: ${err.message}`,
-        "automation"
-      );
-      return [];
-    }
-
-    // Filter for S3 buckets only and map to AWS-style property names
-    const buckets = allResources
-      .filter((resource) => resource.type === "s3")
-      .map((bucket) => ({
-        Name: bucket.name,
-        CreationDate: bucket.createdAt,
-      }));
-    return buckets;
-  } catch (error) {
-    addLog("error", `Failed to list buckets: ${error.message}`, "automation");
-    return [];
-  }
-}
-
-async function listBucketContents(projectName, bucketName) {
-  try {
-    let command = `./list_bucket_contents.sh ${projectName} dev`;
-
-    if (bucketName) {
-      command += ` ${bucketName}`;
-    }
-
-    const { stdout, stderr } = await execAsync(command, {
-      cwd: "/app/scripts/shell",
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog("warn", `Bucket listing warning: ${stderr}`, "automation");
-    }
-
-    // Parse the output as JSON
-    let contents = [];
-    try {
-      // Debug: Log the raw output
-      addLog("info", `Bucket listing raw output: "${stdout}"`, "automation");
-
-      contents = JSON.parse(stdout);
-      if (!Array.isArray(contents)) {
-        addLog(
-          "error",
-          "Bucket listing output is not a JSON array",
-          "automation"
-        );
-        contents = [];
-      }
-    } catch (err) {
-      addLog(
-        "error",
-        `Failed to parse bucket listing JSON: ${err.message}. Raw output: "${stdout}"`,
-        "automation"
-      );
-      contents = [];
-    }
-
-    return contents;
-  } catch (error) {
-    addLog(
-      "error",
-      `Failed to list bucket contents: ${error.message}`,
-      "automation"
-    );
-    return [];
-  }
-}
-
-async function listDynamoDBTables(projectName) {
-  try {
-    let command = `./list_dynamodb_tables.sh ${projectName} --all`;
-
-    const { stdout, stderr } = await execAsync(command, {
-      cwd: "/app/scripts/shell",
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog("warn", `DynamoDB table listing warning: ${stderr}`, "automation");
-    }
-
-    // Parse the output as JSON array
-    let tables = [];
-    try {
-      tables = JSON.parse(stdout);
-      if (!Array.isArray(tables)) {
-        addLog(
-          "error",
-          "DynamoDB table listing output is not a JSON array",
-          "automation"
-        );
-        tables = [];
-      }
-    } catch (err) {
-      addLog(
-        "error",
-        `Failed to parse DynamoDB table listing JSON: ${err.message}`,
-        "automation"
-      );
-      tables = [];
-    }
-
-    return tables;
-  } catch (error) {
-    addLog(
-      "error",
-      `Failed to list DynamoDB tables: ${error.message}`,
-      "automation"
-    );
-    return [];
-  }
-}
-
-async function scanDynamoDBTable(projectName, tableName, limit = 100) {
-  try {
-    let command = `./scan_dynamodb_table.sh ${projectName} ${tableName} ${limit}`;
-
-    const { stdout, stderr } = await execAsync(command, {
-      cwd: "/app/scripts/shell",
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog("warn", `DynamoDB scan warning: ${stderr}`, "automation");
-    }
-
-    // Parse the output as JSON
-    let result = { items: [], count: 0, scannedCount: 0 };
-    try {
-      result = JSON.parse(stdout);
-    } catch (err) {
-      addLog(
-        "error",
-        `Failed to parse DynamoDB scan JSON: ${err.message}`,
-        "automation"
-      );
-    }
-
-    return result;
-  } catch (error) {
-    addLog(
-      "error",
-      `Failed to scan DynamoDB table: ${error.message}`,
-      "automation"
-    );
-    return { items: [], count: 0, scannedCount: 0 };
-  }
-}
-
-async function queryDynamoDBTable(
-  projectName,
-  tableName,
-  partitionKey,
-  partitionValue,
-  sortKey,
-  sortValue,
-  limit = 100,
-  indexName = ""
-) {
-  try {
-    let command = `./query_dynamodb_table.sh ${projectName} ${tableName} "${partitionKey}" "${partitionValue}" "${
-      sortKey || ""
-    }" "${sortValue || ""}" ${limit} "${indexName || ""}"`;
-
-    const { stdout, stderr } = await execAsync(command, {
-      cwd: "/app/scripts/shell",
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog("warn", `DynamoDB query warning: ${stderr}`, "automation");
-    }
-
-    // Parse the output as JSON
-    let result = { items: [], count: 0, scannedCount: 0 };
-    try {
-      result = JSON.parse(stdout);
-    } catch (err) {
-      addLog(
-        "error",
-        `Failed to parse DynamoDB query JSON: ${err.message}`,
-        "automation"
-      );
-    }
-
-    return result;
-  } catch (error) {
-    addLog(
-      "error",
-      `Failed to query DynamoDB table: ${error.message}`,
-      "automation"
-    );
-    return { items: [], count: 0, scannedCount: 0 };
-  }
-}
-
-async function getDynamoDBTableSchema(projectName, tableName) {
-  try {
-    let command = `aws dynamodb describe-table --table-name "${tableName}" --endpoint-url "${internalEndpoint}" --region "${awsRegion}"`;
-
-    const { stdout, stderr } = await execAsync(command, {
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog(
-        "warn",
-        `DynamoDB describe-table warning: ${stderr}`,
-        "automation"
-      );
-    }
-
-    // Parse the output as JSON
-    let result = null;
-    try {
-      result = JSON.parse(stdout);
-    } catch (err) {
-      addLog(
-        "error",
-        `Failed to parse DynamoDB describe-table JSON: ${err.message}`,
-        "automation"
-      );
-      return null;
-    }
-
-    return result;
-  } catch (error) {
-    addLog(
-      "error",
-      `Failed to get DynamoDB table schema: ${error.message}`,
-      "automation"
-    );
-    return null;
-  }
-}
-
-// API Routes
-
-// Health check
-app.get("/health", (req, res) => {
-  res.json({
-    status: "ok",
-    timestamp: new Date().toISOString(),
-    service: "LocalCloud Kit API",
-    vendor: "CloudStack Solutions",
-    version: "0.5.3",
-  });
-});
-
-// Batched dashboard data (single round-trip for dashboard load)
-app.get("/dashboard", async (req, res) => {
-  try {
-    const projectConfigRow = db
-      .prepare(
-        `SELECT p.name as projectName
-         FROM user_profile u
-         LEFT JOIN projects p ON u.active_project_id = p.id
-         WHERE u.id = 1`
-      )
-      .get();
-    const projectConfig = {
-      projectName: projectConfigRow?.projectName || "default",
-      awsEndpoint: userEndpoint,
-      awsRegion,
-    };
-
-    const localstackStatusForDashboard = {
-      ...localstackStatus,
-      endpoint: userEndpoint,
-    };
-
-    const MAILPIT_URL =
-      process.env.MAILPIT_INTERNAL_URL || "http://mailpit:8025";
-
-    const [mailpitResult, resources, cacheResult] = await Promise.all([
-      axios
-        .get(`${MAILPIT_URL}/api/v1/info`, { timeout: 3000 })
-        .then((r) => ({
-          total: r.data.Messages ?? 0,
-          unread: r.data.Unread ?? 0,
-          status: "healthy",
-        }))
-        .catch(() => ({ total: 0, unread: 0, status: "unavailable" })),
-      listResources(projectConfig.projectName),
-      execAsync(`/bin/sh ${path.join("/app/scripts/shell", "list_cache.sh")}`)
-        .then(({ stdout }) => JSON.parse(stdout))
-        .catch(() => ({ status: "unknown", info: null })),
-    ]);
-
-    const redisStatus =
-      cacheResult.status === "running" ? "running" : "stopped";
-    const resourcesWithExtras = [...resources];
-    resourcesWithExtras.push({
-      id: "cache-redis",
-      name: "Redis Cache",
-      type: "cache",
-      status: redisStatus === "running" ? "active" : "error",
-      environment: "local",
-      project: projectConfig.projectName,
-      createdAt: new Date().toISOString(),
-      details: {
-        info: cacheResult.info,
-        status: cacheResult.status,
-      },
-    });
-    resourcesWithExtras.push({
-      id: "mailpit-inbox",
-      name: "Inbox",
-      type: "mailpit",
-      status: mailpitResult.status === "healthy" ? "active" : "error",
-      environment: "local",
-      project: projectConfig.projectName,
-      createdAt: new Date().toISOString(),
-      details: {
-        total: mailpitResult.total,
-        unread: mailpitResult.unread,
-        status: mailpitResult.status,
-      },
-    });
-
-    res.json({
-      success: true,
-      data: {
-        localstackStatus: localstackStatusForDashboard,
-        projectConfig,
-        mailpit: mailpitResult,
-        resources: resourcesWithExtras,
-        redis: {
-          status: redisStatus,
-          info: cacheResult.info,
-        },
-      },
-    });
-  } catch (err) {
-    addLog("error", `Dashboard aggregation failed: ${err.message}`, "api");
-    res.status(500).json({
-      success: false,
-      error: err.message || "Failed to load dashboard data",
-    });
-  }
-});
-
-// LocalStack Management
-app.get("/localstack/status", (req, res) => {
-  // Return user-friendly endpoint for GUI display
-  const userFriendlyStatus = {
-    ...localstackStatus,
-    endpoint: userEndpoint,
-  };
-  res.json({ success: true, data: userFriendlyStatus });
-});
-
-app.get("/localstack/logs", (req, res) => {
-  res.json({ success: true, data: logs });
-});
-
-// Resource Management
-app.get("/resources/list", async (req, res) => {
-  const { projectName } = req.query;
-  const resources = await listResources(projectName);
-  res.json({ success: true, data: resources });
-});
-
-app.post("/resources/create", async (req, res) => {
-  const result = await createResources(req.body);
-  res.json(result);
-});
-
-app.post("/resources/create-single", async (req, res) => {
-  try {
-    const { projectName, resourceType, ...config } = req.body;
-
-    if (!projectName || !resourceType) {
-      return res.status(400).json({
-        success: false,
-        error: "projectName and resourceType are required",
-      });
-    }
-
-    const result = await createSingleResource(
-      projectName,
-      resourceType,
-      config
-    );
-    res.json({
-      success: true,
-      message: `${resourceType} resource created successfully`,
-      data: result,
-    });
-  } catch (error) {
-    res.status(500).json({
-      success: false,
-      error: error.message,
-    });
-  }
-});
-
-app.post("/resources/destroy", async (req, res) => {
-  const result = await destroyResources(req.body);
-  res.json(result);
-});
-
-app.post("/resources/destroy-single", async (req, res) => {
-  try {
-    const { projectName, resourceType, resourceName } = req.body;
-
-    if (!projectName || !resourceType) {
-      return res.status(400).json({
-        success: false,
-        error: "projectName and resourceType are required",
-      });
-    }
-
-    const result = await destroySingleResource(
-      projectName,
-      resourceType,
-      resourceName
-    );
-    res.json(result);
-  } catch (error) {
-    res.status(500).json({
-      success: false,
-      error: error.message,
-    });
-  }
-});
-
-app.get("/resources/status", async (req, res) => {
-  const { projectName } = req.query;
-  const resources = await listResources(projectName);
-  res.json({ success: true, data: resources });
-});
-
-// Configuration Management
-app.get("/config/project", (req, res) => {
-  const row = db
-    .prepare(
-      `SELECT p.name as projectName
-       FROM user_profile u
-       LEFT JOIN projects p ON u.active_project_id = p.id
-       WHERE u.id = 1`
-    )
-    .get();
-  res.json({
-    success: true,
-    data: {
-      projectName: row?.projectName || "default",
-      awsEndpoint: userEndpoint,
-      awsRegion,
-    },
-  });
-});
-
-// Profile
-app.get("/profile", (req, res) => {
-  const profile = db
-    .prepare(
-      `SELECT u.*, p.name as active_project_name, p.label as active_project_label
-       FROM user_profile u
-       LEFT JOIN projects p ON u.active_project_id = p.id
-       WHERE u.id = 1`
-    )
-    .get();
-  res.json({ success: true, data: profile });
-});
-
-app.put("/profile", (req, res) => {
-  const { preferred_language, highlight_theme, display_name, active_project_id } = req.body;
-  const sets = [];
-  const vals = [];
-  if (preferred_language !== undefined) { sets.push("preferred_language = ?"); vals.push(preferred_language); }
-  if (highlight_theme !== undefined) { sets.push("highlight_theme = ?"); vals.push(highlight_theme); }
-  if (display_name !== undefined) { sets.push("display_name = ?"); vals.push(display_name); }
-  if (active_project_id !== undefined) { sets.push("active_project_id = ?"); vals.push(active_project_id); }
-  if (sets.length === 0) return res.status(400).json({ success: false, error: "No fields to update" });
-  sets.push("updated_at = CURRENT_TIMESTAMP");
-  db.prepare(`UPDATE user_profile SET ${sets.join(", ")} WHERE id = 1`).run(...vals);
-  const profile = db
-    .prepare(
-      `SELECT u.*, p.name as active_project_name, p.label as active_project_label
-       FROM user_profile u
-       LEFT JOIN projects p ON u.active_project_id = p.id
-       WHERE u.id = 1`
-    )
-    .get();
-  res.json({ success: true, data: profile });
-});
-
-// Projects
-app.get("/projects", (req, res) => {
-  const projects = db.prepare("SELECT * FROM projects ORDER BY created_at ASC").all();
-  res.json({ success: true, data: projects });
-});
-
-app.post("/projects", (req, res) => {
-  const { name, label, description } = req.body;
-  if (!name || !label) return res.status(400).json({ success: false, error: "name and label are required" });
-  try {
-    const result = db
-      .prepare("INSERT INTO projects (name, label, description) VALUES (?, ?, ?)")
-      .run(name, label, description || null);
-    const project = db.prepare("SELECT * FROM projects WHERE id = ?").get(result.lastInsertRowid);
-    res.json({ success: true, data: project });
-  } catch {
-    res.status(400).json({ success: false, error: "Project name already exists" });
-  }
-});
-
-app.put("/projects/:id", (req, res) => {
-  const { label, description } = req.body;
-  db.prepare("UPDATE projects SET label = COALESCE(?, label), description = COALESCE(?, description) WHERE id = ?")
-    .run(label || null, description !== undefined ? description : null, req.params.id);
-  const project = db.prepare("SELECT * FROM projects WHERE id = ?").get(req.params.id);
-  if (!project) return res.status(404).json({ success: false, error: "Project not found" });
-  res.json({ success: true, data: project });
-});
-
-app.delete("/projects/:id", (req, res) => {
-  const profile = db.prepare("SELECT active_project_id FROM user_profile WHERE id = 1").get();
-  if (profile?.active_project_id == req.params.id) {
-    return res.status(400).json({ success: false, error: "Cannot delete the active project. Switch projects first." });
-  }
-  db.prepare("DELETE FROM projects WHERE id = ?").run(req.params.id);
-  res.json({ success: true });
-});
-
-// Saved Configs
-app.get("/saved-configs", (req, res) => {
-  const { project_id, type } = req.query;
-  let query = "SELECT * FROM saved_configs WHERE 1=1";
-  const params = [];
-  if (project_id) { query += " AND project_id = ?"; params.push(project_id); }
-  if (type) { query += " AND resource_type = ?"; params.push(type); }
-  query += " ORDER BY created_at DESC";
-  const rows = db.prepare(query).all(...params);
-  const configs = rows.map((r) => ({ ...r, config: JSON.parse(r.config_json) }));
-  res.json({ success: true, data: configs });
-});
-
-app.post("/saved-configs", (req, res) => {
-  const { project_id, name, resource_type, config } = req.body;
-  if (!project_id || !name || !resource_type || !config) {
-    return res.status(400).json({ success: false, error: "project_id, name, resource_type, and config are required" });
-  }
-  const result = db
-    .prepare("INSERT INTO saved_configs (project_id, name, resource_type, config_json) VALUES (?, ?, ?, ?)")
-    .run(project_id, name, resource_type, JSON.stringify(config));
-  const row = db.prepare("SELECT * FROM saved_configs WHERE id = ?").get(result.lastInsertRowid);
-  res.json({ success: true, data: { ...row, config: JSON.parse(row.config_json) } });
-});
-
-app.put("/saved-configs/:id", (req, res) => {
-  const { name, config } = req.body;
-  db.prepare(
-    "UPDATE saved_configs SET name = COALESCE(?, name), config_json = COALESCE(?, config_json), updated_at = CURRENT_TIMESTAMP WHERE id = ?"
-  ).run(name || null, config ? JSON.stringify(config) : null, req.params.id);
-  const row = db.prepare("SELECT * FROM saved_configs WHERE id = ?").get(req.params.id);
-  if (!row) return res.status(404).json({ success: false, error: "Saved config not found" });
-  res.json({ success: true, data: { ...row, config: JSON.parse(row.config_json) } });
-});
-
-app.delete("/saved-configs/:id", (req, res) => {
-  db.prepare("DELETE FROM saved_configs WHERE id = ?").run(req.params.id);
-  res.json({ success: true });
-});
-
-app.get("/config/templates", (req, res) => {
-  const templates = [
-    {
-      id: "basic",
-      name: "Basic Setup",
-      description: "S3 bucket and DynamoDB table for basic storage needs",
-      resources: { s3: true, dynamodb: true, lambda: false, apigateway: false },
-    },
-    {
-      id: "serverless",
-      name: "Serverless Application",
-      description: "Complete serverless stack with Lambda and API Gateway",
-      resources: { s3: true, dynamodb: true, lambda: true, apigateway: true },
-    },
-    {
-      id: "storage",
-      name: "Storage Only",
-      description: "S3 bucket for file storage",
-      resources: {
-        s3: true,
-        dynamodb: false,
-        lambda: false,
-        apigateway: false,
-      },
-    },
-    {
-      id: "database",
-      name: "Database Only",
-      description: "DynamoDB table for data storage",
-      resources: {
-        s3: false,
-        dynamodb: true,
-        lambda: false,
-        apigateway: false,
-      },
-    },
-    {
-      id: "api",
-      name: "API Only",
-      description: "API Gateway with Lambda function",
-      resources: { s3: false, dynamodb: false, lambda: true, apigateway: true },
-    },
-  ];
-  res.json({ success: true, data: templates });
-});
-
-// S3 Bucket Management
-app.get("/s3/buckets", async (req, res) => {
-  const { projectName } = req.query;
-  const buckets = await listAllBuckets(projectName);
-  res.json({ success: true, data: buckets });
-});
-
-app.get("/s3/bucket/:bucketName/contents", async (req, res) => {
-  const { projectName } = req.query;
-  const { bucketName } = req.params;
-  const contents = await listBucketContents(projectName, bucketName);
-  res.json({ success: true, data: contents });
-});
-
-app.get("/s3/bucket/:bucketName/object/*", async (req, res) => {
-  try {
-    const { projectName } = req.query;
-    const { bucketName } = req.params;
-    const objectKey = req.params[0]; // Get the wildcard parameter
-
-    if (!projectName || !bucketName || !objectKey) {
-      return res.status(400).json({
-        success: false,
-        error: "projectName, bucketName, and objectKey are required",
-      });
-    }
-
-    const command = `/bin/sh /app/scripts/shell/download_s3_object.sh '${projectName}' '${bucketName}' '${objectKey}'`;
-    const { stdout, stderr } = await execAsync(command, {
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    // Debug: Log the raw output
-    console.log("[S3 Download] Raw stdout:", JSON.stringify(stdout));
-    console.log("[S3 Download] Raw stderr:", JSON.stringify(stderr));
-    addLog("info", `[S3 Download] Raw stdout: ${JSON.stringify(stdout)}`);
-    addLog("info", `[S3 Download] Raw stderr: ${JSON.stringify(stderr)}`);
-
-    // Parse metadata from stderr (it's output as a comment)
-    let metadata = {};
-    if (stderr) {
-      const metadataMatch = stderr.match(/<!--METADATA:(.*?)-->/);
-      if (metadataMatch) {
-        try {
-          metadata = JSON.parse(metadataMatch[1]);
-        } catch (e) {
-          addLog(
-            "warn",
-            `Failed to parse object metadata: ${e.message}`,
-            "automation"
-          );
-        }
-      }
-    }
-
-    // Determine if the file is binary or text
-    const contentType = metadata.ContentType || "";
-    const isBinary =
-      contentType.startsWith("image/") ||
-      contentType === "application/pdf" ||
-      contentType.startsWith("application/zip") ||
-      contentType.startsWith("application/octet-stream");
-
-    let fileContent;
-    if (isBinary) {
-      // The shell script now outputs binary files as base64, so use stdout directly
-      fileContent = stdout;
-    } else {
-      fileContent = stdout;
-    }
-
-    addLog(
-      "success",
-      `Object downloaded: ${objectKey} from bucket ${bucketName}`,
-      "automation"
-    );
-
-    res.json({
-      success: true,
-      data: {
-        content: fileContent,
-        metadata: metadata,
-      },
-    });
-  } catch (error) {
-    addLog(
-      "error",
-      `Failed to download object: ${error.message}`,
-      "automation"
-    );
-    res.status(500).json({ success: false, error: error.message });
-  }
-});
-
-// Multipart file upload endpoint (recommended)
-app.post(
-  "/s3/bucket/:bucketName/upload-multipart",
-  upload.single("file"),
-  async (req, res) => {
-    try {
-      const { projectName } = req.query;
-      const { bucketName } = req.params;
-      const { objectKey } = req.body;
-      const file = req.file;
-
-      if (!projectName || !bucketName || !objectKey || !file) {
-        return res.status(400).json({
-          success: false,
-          error: "projectName, bucketName, objectKey, and file are required",
-        });
-      }
-
-      addLog(
-        "info",
-        `Uploading file: ${objectKey} (${(file.size / 1024).toFixed(
-          2
-        )} KB) to bucket ${bucketName}`,
-        "automation"
-      );
-
-      const command = `/bin/sh /app/scripts/shell/upload_s3_object.sh '${projectName}' '${bucketName}' '${objectKey}' '${file.path}'`;
-      const { stdout, stderr } = await execAsync(command, {
-        env: {
-          ...process.env,
-          AWS_ENDPOINT_URL: internalEndpoint,
-          AWS_DEFAULT_REGION: awsRegion,
-        },
-      });
-
-      // Clean up temp file
-      try {
-        fs.unlinkSync(file.path);
-      } catch (cleanupError) {
-        console.warn("Failed to cleanup temp file:", cleanupError);
-      }
-
-      if (stderr) {
-        addLog("warn", `S3 upload warning: ${stderr}`, "automation");
-      }
-
-      addLog(
-        "success",
-        `Object uploaded: ${objectKey} (${(file.size / 1024).toFixed(
-          2
-        )} KB) to bucket ${bucketName}`,
-        "automation"
-      );
-
-      res.json({
-        success: true,
-        message: `File uploaded successfully`,
-        size: file.size,
-        filename: objectKey,
-      });
-    } catch (error) {
-      // Clean up file on error
-      if (req.file && req.file.path) {
-        try {
-          fs.unlinkSync(req.file.path);
-        } catch (cleanupError) {
-          console.warn("Failed to cleanup temp file on error:", cleanupError);
-        }
-      }
-      addLog(
-        "error",
-        `Failed to upload object: ${error.message}`,
-        "automation"
-      );
-      res.status(500).json({ success: false, error: error.message });
-    }
-  }
-);
-
-// Legacy JSON-based upload endpoint (kept for backward compatibility)
-app.post("/s3/bucket/:bucketName/upload", async (req, res) => {
-  try {
-    const { projectName } = req.query;
-    const { bucketName } = req.params;
-    const { objectKey, content } = req.body;
-
-    if (!projectName || !bucketName || !objectKey || content === undefined) {
-      return res.status(400).json({
-        success: false,
-        error: "projectName, bucketName, objectKey, and content are required",
-      });
-    }
-
-    // Create a temporary file with the content
-    const tempFile = `/tmp/${objectKey.replace(/[^a-zA-Z0-9.-]/g, "_")}`;
-
-    // Check if content is base64 encoded (for binary files)
-    const isBase64 =
-      /^[A-Za-z0-9+/]*={0,2}$/.test(content) && content.length > 0;
-
-    if (isBase64) {
-      // Decode base64 content for binary files
-      const buffer = Buffer.from(content, "base64");
-      fs.writeFileSync(tempFile, buffer);
-    } else {
-      // Write as text for text files
-      fs.writeFileSync(tempFile, content);
-    }
-
-    const command = `/bin/sh /app/scripts/shell/upload_s3_object.sh '${projectName}' '${bucketName}' '${objectKey}' '${tempFile}'`;
-    const { stdout, stderr } = await execAsync(command, {
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    // Clean up temp file
-    try {
-      fs.unlinkSync(tempFile);
-    } catch (cleanupError) {
-      console.warn("Failed to cleanup temp file:", cleanupError);
-    }
-
-    if (stderr) {
-      addLog("warn", `S3 upload warning: ${stderr}`, "automation");
-    }
-
-    addLog(
-      "success",
-      `Object uploaded: ${objectKey} to bucket ${bucketName}`,
-      "automation"
-    );
-
-    res.json({ success: true, message: stdout });
-  } catch (error) {
-    addLog("error", `Failed to upload object: ${error.message}`, "automation");
-    res.status(500).json({ success: false, error: error.message });
-  }
-});
-
-app.delete("/s3/bucket/:bucketName/object/*", async (req, res) => {
-  try {
-    const { projectName } = req.query;
-    const { bucketName } = req.params;
-    const objectKey = req.params[0]; // Get the wildcard parameter
-
-    if (!projectName || !bucketName || !objectKey) {
-      return res.status(400).json({
-        success: false,
-        error: "projectName, bucketName, and objectKey are required",
-      });
-    }
-
-    const command = `/bin/sh /app/scripts/shell/delete_s3_object.sh '${projectName}' '${bucketName}' '${objectKey}'`;
-    const { stdout, stderr } = await execAsync(command, {
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog("warn", `S3 delete-object warning: ${stderr}`, "automation");
-    }
-
-    addLog(
-      "success",
-      `Object deleted: ${objectKey} from bucket ${bucketName}`,
-      "automation"
-    );
-
-    res.json({ success: true, message: stdout });
-  } catch (error) {
-    addLog("error", `Failed to delete object: ${error.message}`, "automation");
-    res.status(500).json({ success: false, error: error.message });
-  }
-});
-
-// DynamoDB Table Management
-app.get("/dynamodb/tables", async (req, res) => {
-  try {
-    const { projectName } = req.query;
-    const tables = await listDynamoDBTables(projectName);
-    res.json({ success: true, data: tables });
-  } catch (error) {
-    res.status(500).json({
-      success: false,
-      error: error.message,
-    });
-  }
-});
-
-app.get("/dynamodb/table/:tableName/scan", async (req, res) => {
-  try {
-    const { projectName, limit = "100" } = req.query;
-    const { tableName } = req.params;
-    const result = await scanDynamoDBTable(
-      projectName,
-      tableName,
-      parseInt(limit)
-    );
-    res.json({ success: true, data: result });
-  } catch (error) {
-    res.status(500).json({
-      success: false,
-      error: error.message,
-    });
-  }
-});
-
-app.get("/dynamodb/table/:tableName/query", async (req, res) => {
-  try {
-    const {
-      projectName,
-      partitionKey,
-      partitionValue,
-      sortKey,
-      sortValue,
-      limit = "100",
-      indexName,
-    } = req.query;
-    const { tableName } = req.params;
-
-    const result = await queryDynamoDBTable(
-      projectName,
-      tableName,
-      partitionKey,
-      partitionValue,
-      sortKey,
-      sortValue,
-      parseInt(limit),
-      indexName
-    );
-    res.json({ success: true, data: result });
-  } catch (error) {
-    res.status(500).json({
-      success: false,
-      error: error.message,
-    });
-  }
-});
-
-app.post("/dynamodb/table/:tableName/item", async (req, res) => {
-  try {
-    const { projectName } = req.body;
-    const { tableName } = req.params;
-    const { item } = req.body;
-    if (!projectName || !tableName || !item) {
-      return res.status(400).json({
-        success: false,
-        error: "projectName, tableName, and item are required",
-      });
-    }
-    // Stringify item for shell script
-    const itemJson = JSON.stringify(item).replace(/'/g, "'''");
-    const command = `/bin/sh /app/scripts/shell/put_dynamodb_item.sh '${projectName}' '${tableName}' '${itemJson}'`;
-    const { stdout, stderr } = await execAsync(command, {
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-    if (stderr) {
-      addLog("warn", `DynamoDB put-item warning: ${stderr}`, "automation");
-    }
-    addLog(
-      "success",
-      `Item added to DynamoDB table ${tableName}`,
-      "automation"
-    );
-    res.json({ success: true });
-  } catch (error) {
-    addLog(
-      "error",
-      `Failed to add item to DynamoDB: ${error.message}`,
-      "automation"
-    );
-    res.status(500).json({ success: false, error: error.message });
-  }
-});
-
-app.delete("/dynamodb/table/:tableName/item", async (req, res) => {
-  try {
-    const { projectName } = req.body;
-    const { tableName } = req.params;
-    const { partitionKey, partitionValue, sortKey, sortValue } = req.body;
-
-    if (!projectName || !tableName || !partitionKey || !partitionValue) {
-      return res.status(400).json({
-        success: false,
-        error:
-          "projectName, tableName, partitionKey, and partitionValue are required",
-      });
-    }
-
-    let command = `/bin/sh /app/scripts/shell/delete_dynamodb_item.sh '${projectName}' '${tableName}' '${partitionKey}' '${partitionValue}'`;
-
-    if (sortKey && sortValue) {
-      command += ` '${sortKey}' '${sortValue}'`;
-    }
-
-    const { stdout, stderr } = await execAsync(command, {
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog("warn", `DynamoDB delete-item warning: ${stderr}`, "automation");
-    }
-
-    addLog(
-      "success",
-      `Item deleted from DynamoDB table ${tableName}`,
-      "automation"
-    );
-
-    res.json({ success: true, message: stdout });
-  } catch (error) {
-    addLog(
-      "error",
-      `Failed to delete item from DynamoDB: ${error.message}`,
-      "automation"
-    );
-    res.status(500).json({ success: false, error: error.message });
-  }
-});
-
-app.get("/dynamodb/table/:tableName/schema", async (req, res) => {
-  try {
-    const { projectName } = req.query;
-    const { tableName } = req.params;
-
-    if (!projectName || !tableName) {
-      return res.status(400).json({
-        success: false,
-        error: "projectName and tableName are required",
-      });
-    }
-
-    const result = await getDynamoDBTableSchema(projectName, tableName);
-    if (!result) {
-      return res.status(404).json({
-        success: false,
-        error: "Table not found or failed to get schema",
-      });
-    }
-
-    res.json({ success: true, data: result });
-  } catch (error) {
-    res.status(500).json({
-      success: false,
-      error: error.message,
-    });
-  }
-});
-
-// Redis Cache Management
-const pathToShell = (script) => path.join("/app/scripts/shell", script);
-
-app.get("/cache/status", async (req, res) => {
-  try {
-    const { stdout } = await execAsync(
-      `/bin/sh ${pathToShell("list_cache.sh")}`
-    );
-    res.json(JSON.parse(stdout));
-  } catch (error) {
-    res.status(500).json({ success: false, error: error.message });
-  }
-});
-
-app.post("/cache/set", async (req, res) => {
-  const { key, value } = req.body;
-  if (!key || value === undefined) {
-    return res
-      .status(400)
-      .json({ success: false, error: "Key and value required" });
-  }
-  try {
-    const { stdout } = await execAsync(
-      `/bin/sh ${pathToShell("cache_set.sh")} '${key}' '${value}'`
-    );
-    res.json(JSON.parse(stdout));
-  } catch (error) {
-    res.status(500).json({ success: false, error: error.message });
-  }
-});
-
-app.get("/cache/get", async (req, res) => {
-  const { key } = req.query;
-  if (!key) {
-    return res.status(400).json({ success: false, error: "Key required" });
-  }
-  try {
-    const { stdout } = await execAsync(
-      `/bin/sh ${pathToShell("cache_get.sh")} '${key}'`
-    );
-    res.json(JSON.parse(stdout));
-  } catch (error) {
-    res.status(500).json({ success: false, error: error.message });
-  }
-});
-
-app.delete("/cache/del", async (req, res) => {
-  const { key } = req.body;
-  if (!key) {
-    return res.status(400).json({ success: false, error: "Key required" });
-  }
-  try {
-    const { stdout } = await execAsync(
-      `/bin/sh ${pathToShell("cache_del.sh")} '${key}'`
-    );
-    res.json(JSON.parse(stdout));
-  } catch (error) {
-    res.status(500).json({ success: false, error: error.message });
-  }
-});
-
-app.post("/cache/flush", async (req, res) => {
-  try {
-    const { stdout } = await execAsync(
-      `/bin/sh ${pathToShell("cache_flush.sh")}`
-    );
-    res.json(JSON.parse(stdout));
-  } catch (error) {
-    res.status(500).json({ success: false, error: error.message });
-  }
-});
-
-app.get("/cache/keys", async (req, res) => {
-  try {
-    const { stdout } = await execAsync(
-      `/bin/sh ${pathToShell("list_cache_keys.sh")}`
-    );
-    res.json(JSON.parse(stdout));
-  } catch (error) {
-    res.status(500).json({ success: false, error: error.message });
-  }
-});
+// Routes
+app.use(healthRouter);
+app.use(localstackRouter);
+app.use(resourcesRouter);
+app.use(configRouter);
+app.use(profileRouter);
+app.use(projectsRouter);
+app.use(savedConfigsRouter);
+app.use(s3Router);
+app.use(dynamodbRouter);
+app.use(cacheRouter);
+app.use(mailpitRouter);
+app.use(secretsRouter);
+app.use(postgresRouter);
+app.use(keycloakRouter);
 
 // Socket.IO connection handling
 io.on("connection", (socket) => {
   addLog("info", "Client connected", "gui");
-
   socket.on("disconnect", () => {
     addLog("info", "Client disconnected", "gui");
   });
@@ -1698,452 +73,6 @@ io.on("connection", (socket) => {
 cron.schedule("*/30 * * * * *", () => {
   checkLocalStackStatus();
 });
-
-// Secrets Manager API endpoints
-app.get("/secrets", async (req, res) => {
-  try {
-    const { maxResults = "100", nextToken = "" } = req.query;
-
-    const command = `/bin/sh /app/scripts/shell/list_secrets.sh ${maxResults} ${nextToken}`;
-
-    const { stdout, stderr } = await execAsync(command, {
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog("warn", `Secrets list warning: ${stderr}`, "automation");
-    }
-
-    const secrets = JSON.parse(stdout);
-    res.json({ success: true, data: secrets });
-  } catch (error) {
-    addLog("error", `Failed to list secrets: ${error.message}`, "automation");
-    res.status(500).json({
-      success: false,
-      error: "Failed to list secrets",
-      message: error.message,
-    });
-  }
-});
-
-app.post("/secrets", async (req, res) => {
-  try {
-    const { secretName, secretValue, description, tags, kmsKeyId } = req.body;
-
-    if (!secretName || !secretValue) {
-      return res.status(400).json({
-        success: false,
-        error: "Secret name and value are required",
-      });
-    }
-
-    // Build command arguments
-    let command = `/bin/sh /app/scripts/shell/create_secret.sh "${secretName}" "${secretValue}"`;
-
-    if (description) {
-      command += ` "${description}"`;
-    } else {
-      command += ` ""`;
-    }
-
-    if (tags && Object.keys(tags).length > 0) {
-      const tagString = Object.entries(tags)
-        .map(([key, value]) => `Key=${key},Value=${value}`)
-        .join(" ");
-      command += ` "${tagString}"`;
-    } else {
-      command += ` ""`;
-    }
-
-    if (kmsKeyId) {
-      command += ` "${kmsKeyId}"`;
-    } else {
-      command += ` ""`;
-    }
-
-    const { stdout, stderr } = await execAsync(command, {
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog("warn", `Secret creation warning: ${stderr}`, "automation");
-    }
-
-    addLog(
-      "success",
-      `Secret ${secretName} created successfully`,
-      "automation"
-    );
-
-    res.json({
-      success: true,
-      data: {
-        secretName,
-        message: stdout.trim(),
-      },
-    });
-  } catch (error) {
-    addLog("error", `Failed to create secret: ${error.message}`, "automation");
-    res.status(500).json({
-      success: false,
-      error: "Failed to create secret",
-      message: error.message,
-    });
-  }
-});
-
-app.get("/secrets/:secretName", async (req, res) => {
-  try {
-    const { secretName } = req.params;
-    const {
-      versionId = "",
-      versionStage = "AWSCURRENT",
-      includeValue = "false",
-    } = req.query;
-
-    if (!secretName) {
-      return res.status(400).json({
-        success: false,
-        error: "Secret name is required",
-      });
-    }
-
-    // Build command arguments
-    let command = `/bin/sh /app/scripts/shell/get_secret.sh "${secretName}"`;
-
-    if (versionId) {
-      command += ` "${versionId}"`;
-    } else {
-      command += ` ""`;
-    }
-
-    command += ` "${versionStage}"`;
-
-    const { stdout, stderr } = await execAsync(command, {
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog("warn", `Get secret warning: ${stderr}`, "automation");
-    }
-
-    const secretData = JSON.parse(stdout);
-
-    // If includeValue is false, mask the secret value
-    if (includeValue !== "true" && secretData.SecretString) {
-      secretData.SecretString = "***MASKED***";
-    }
-
-    res.json({
-      success: true,
-      data: secretData,
-    });
-  } catch (error) {
-    addLog("error", `Failed to get secret: ${error.message}`, "automation");
-    res.status(500).json({
-      success: false,
-      error: "Failed to get secret",
-      message: error.message,
-    });
-  }
-});
-
-app.put("/secrets/:secretName", async (req, res) => {
-  try {
-    const { secretName } = req.params;
-    const { secretValue, description, tags, kmsKeyId } = req.body;
-
-    if (!secretName) {
-      return res.status(400).json({
-        success: false,
-        error: "Secret name is required",
-      });
-    }
-
-    if (!secretValue) {
-      return res.status(400).json({
-        success: false,
-        error: "Secret value is required",
-      });
-    }
-
-    // For updating, we'll use the AWS CLI directly since we need to handle updates
-    let command = `aws --endpoint-url=${internalEndpoint} --region=${awsRegion} secretsmanager update-secret --secret-id "${secretName}" --secret-string "${secretValue}"`;
-
-    if (description) {
-      command += ` --description "${description}"`;
-    }
-
-    if (kmsKeyId) {
-      command += ` --kms-key-id "${kmsKeyId}"`;
-    }
-
-    const { stdout, stderr } = await execAsync(command, {
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog("warn", `Update secret warning: ${stderr}`, "automation");
-    }
-
-    // If tags are provided, update them separately
-    if (tags && Object.keys(tags).length > 0) {
-      const tagString = Object.entries(tags)
-        .map(([key, value]) => `Key=${key},Value=${value}`)
-        .join(" ");
-
-      const tagCommand = `aws --endpoint-url=${internalEndpoint} --region=${awsRegion} secretsmanager tag-resource --secret-id "${secretName}" --tags ${tagString}`;
-
-      await execAsync(tagCommand, {
-        env: {
-          ...process.env,
-          AWS_ENDPOINT_URL: internalEndpoint,
-          AWS_DEFAULT_REGION: awsRegion,
-        },
-      });
-    }
-
-    addLog(
-      "success",
-      `Secret ${secretName} updated successfully`,
-      "automation"
-    );
-
-    res.json({
-      success: true,
-      data: {
-        secretName,
-        message: "Secret updated successfully",
-      },
-    });
-  } catch (error) {
-    addLog("error", `Failed to update secret: ${error.message}`, "automation");
-    res.status(500).json({
-      success: false,
-      error: "Failed to update secret",
-      message: error.message,
-    });
-  }
-});
-
-app.delete("/secrets/:secretName", async (req, res) => {
-  try {
-    const { secretName } = req.params;
-    const { forceDelete = "false" } = req.query;
-
-    if (!secretName) {
-      return res.status(400).json({
-        success: false,
-        error: "Secret name is required",
-      });
-    }
-
-    const command = `/bin/sh /app/scripts/shell/delete_secret.sh "${secretName}" ${forceDelete}`;
-
-    const { stdout, stderr } = await execAsync(command, {
-      env: {
-        ...process.env,
-        AWS_ENDPOINT_URL: internalEndpoint,
-        AWS_DEFAULT_REGION: awsRegion,
-      },
-    });
-
-    if (stderr) {
-      addLog("warn", `Delete secret warning: ${stderr}`, "automation");
-    }
-
-    addLog(
-      "success",
-      `Secret ${secretName} deleted successfully`,
-      "automation"
-    );
-
-    res.json({
-      success: true,
-      data: {
-        secretName,
-        message: stdout.trim(),
-      },
-    });
-  } catch (error) {
-    addLog("error", `Failed to delete secret: ${error.message}`, "automation");
-    res.status(500).json({
-      success: false,
-      error: "Failed to delete secret",
-      message: error.message,
-    });
-  }
-});
-
-// ─── Mailpit Email Testing ───────────────────────────────────────────────────
-
-const MAILPIT_INTERNAL_URL =
-  process.env.MAILPIT_INTERNAL_URL || "http://mailpit:8025";
-
-// GET /mailpit/stats — total + unread email counts
-app.get("/mailpit/stats", async (req, res) => {
-  try {
-    const response = await axios.get(`${MAILPIT_INTERNAL_URL}/api/v1/info`, {
-      timeout: 3000,
-    });
-    const { Messages: total = 0, Unread: unread = 0 } = response.data;
-    res.json({ success: true, data: { total, unread, status: "healthy" } });
-  } catch (error) {
-    res.json({
-      success: false,
-      data: { total: 0, unread: 0, status: "unavailable" },
-      error: error.message,
-    });
-  }
-});
-
-// GET /mailpit/messages — list recent messages
-app.get("/mailpit/messages", async (req, res) => {
-  try {
-    const limit = req.query.limit || 10;
-    const response = await axios.get(
-      `${MAILPIT_INTERNAL_URL}/api/v1/messages?limit=${limit}`,
-      { timeout: 3000 }
-    );
-    res.json({ success: true, data: response.data });
-  } catch (error) {
-    res.json({ success: false, data: { messages: [], total: 0 }, error: error.message });
-  }
-});
-
-// POST /mailpit/send-test — send a test email via Mailpit SMTP
-app.post("/mailpit/send-test", async (req, res) => {
-  const { from, to, subject, body } = req.body;
-  if (!from || !to || !subject || !body) {
-    return res.status(400).json({ success: false, error: "from, to, subject, and body are required" });
-  }
-  try {
-    const transporter = nodemailer.createTransport({
-      host: "mailpit",
-      port: 1025,
-      secure: false,
-    });
-    await transporter.sendMail({ from, to, subject, text: body });
-    addLog("info", `Test email sent to ${to}`, "mailpit");
-    res.json({ success: true, data: { message: `Email sent to ${to}` } });
-  } catch (error) {
-    addLog("error", `Failed to send test email: ${error.message}`, "mailpit");
-    res.status(500).json({ success: false, error: error.message });
-  }
-});
-
-// DELETE /mailpit/messages — clear all messages
-app.delete("/mailpit/messages", async (req, res) => {
-  try {
-    await axios.delete(`${MAILPIT_INTERNAL_URL}/api/v1/messages`, {
-      timeout: 5000,
-    });
-    addLog("info", "Mailpit messages cleared", "mailpit");
-    res.json({ success: true, data: { message: "All messages deleted" } });
-  } catch (error) {
-    addLog("error", `Failed to clear Mailpit messages: ${error.message}`, "mailpit");
-    res.status(500).json({ success: false, error: error.message });
-  }
-});
-
-// ─── PostgreSQL / pgAdmin ────────────────────────────────────────────────────
-
-const net = require("net");
-
-function checkTcpPort(host, port, timeout = 3000) {
-  return new Promise((resolve) => {
-    const socket = new net.Socket();
-    socket.setTimeout(timeout);
-    socket.connect(port, host, () => {
-      socket.destroy();
-      resolve(true);
-    });
-    socket.on("error", () => {
-      socket.destroy();
-      resolve(false);
-    });
-    socket.on("timeout", () => {
-      socket.destroy();
-      resolve(false);
-    });
-  });
-}
-
-// GET /postgres/status — check PostgreSQL connectivity via TCP
-app.get("/postgres/status", async (req, res) => {
-  try {
-    const reachable = await checkTcpPort("postgres", 5432);
-    res.json({
-      success: true,
-      data: {
-        status: reachable ? "running" : "stopped",
-        host: "localhost",
-        port: 5432,
-        user: "localcloud",
-        database: "localcloud",
-      },
-    });
-  } catch {
-    res.json({ success: true, data: { status: "stopped" } });
-  }
-});
-
-// GET /pgadmin/status — check pgAdmin availability
-app.get("/pgadmin/status", async (req, res) => {
-  try {
-    const response = await axios.get("http://pgadmin:80/misc/ping", {
-      timeout: 3000,
-    });
-    res.json({
-      success: true,
-      data: { status: response.status === 200 ? "running" : "stopped" },
-    });
-  } catch {
-    res.json({ success: true, data: { status: "stopped" } });
-  }
-});
-
-// ─── Keycloak ────────────────────────────────────────────────────────────────
-
-// GET /keycloak/status — check Keycloak health
-app.get("/keycloak/status", async (req, res) => {
-  try {
-    const response = await axios.get("http://keycloak:8080/health/ready", {
-      timeout: 5000,
-    });
-    const healthy = response.data?.status === "UP";
-    res.json({
-      success: true,
-      data: { status: healthy ? "running" : "stopped" },
-    });
-  } catch {
-    // Fall back to TCP check if health endpoint isn't yet available
-    try {
-      const reachable = await checkTcpPort("keycloak", 8080);
-      res.json({ success: true, data: { status: reachable ? "running" : "stopped" } });
-    } catch {
-      res.json({ success: true, data: { status: "stopped" } });
-    }
-  }
-});
-
-// ─────────────────────────────────────────────────────────────────────────────
 
 // Initial status check
 setTimeout(checkLocalStackStatus, 2000);
@@ -2155,11 +84,8 @@ server.listen(PORT, () => {
   console.log(
     `LocalCloud Kit API server running on port ${PORT}`,
     `\nEnvironment: ${process.env.NODE_ENV || "local"}`,
-    `\nLocalStack endpoint: ${
-      process.env.LOCALSTACK_ENDPOINT || internalEndpoint
-    }`
+    `\nLocalStack endpoint: ${process.env.LOCALSTACK_ENDPOINT || internalEndpoint}`
   );
-
   console.log(`
 ╔══════════════════════════════════════════════════════════════╗
 ║                    🚀 LocalCloud Kit                    ║
